% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization_Least_Squares.R
\name{Maximization_Fit}
\alias{Maximization_Fit}
\title{Optimization (Maximization) for Linear Regression Function}
\usage{
Maximization_Fit(p, x, y)
}
\arguments{
\item{p}{Vector of Initial values for searching including sigma}

\item{x}{Vector of Predictors}

\item{y}{Vector of Outcome}
}
\value{
Vector of Intercept and slope (coeffecients_max)
}
\description{
Optimization (Maximization) for Linear Regression Function
}
\examples{
Maximization_Fit(c(0,0,1), c(1,2,3), c(3,5,7))
}
