% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization_Least_Squares.R
\name{Minimization_Fit}
\alias{Minimization_Fit}
\title{Optimization (Minimization) for Linear Regression Function}
\usage{
Minimization_Fit(p, x, y)
}
\arguments{
\item{p}{Vector of Initial values for searching}

\item{x}{Vector of Predictors}

\item{y}{Vector of Outcome}
}
\value{
Vector of Intercept and slope (coeffecients_min)
}
\description{
Optimization (Minimization) for Linear Regression Function
}
\examples{
Minimization_Fit(c(0,0), c(1,2,3), c(3,5,7))
}
