% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Optimization_Data}
\alias{Optimization_Data}
\title{Optimization Sample Data}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{Anxiety_Level}{Predictor to be minimized}
\item{Performance_Score}{Outcome to be maximized}
}
}
\usage{
Optimization_Data
}
\description{
Simulated data for testing minimization and maximization fitting functions.
}
\keyword{datasets}
